Add wallpaper submission popup with image , name and category
